notebooks/requirements.txt
geopandas
cartoframes==1.0b1
carto-print
matplotlib
seaborn
pandas
dask
netCDF4
jupyter
numpy
scipy
sklearn
shapely
fiona
scikit-gstat
pyproj
utm
pysal
pointpats
rpy2
ipywidgets